Fix up migrations

+++ migrations/2018-05-27-102403_create_orders/up.sql
@@ -1,8 +1,8 @@
 -- Your SQL goes here
 CREATE TABLE orders (
     id UUID PRIMARY KEY,
-    price NUMERIC NOT NULL CHECK (price > 0),
-    quantity NUMERIC NOT NULL CHECK (quantity > 0),
+    price INTEGER NOT NULL CHECK (price > 0),
+    quantity INTEGER NOT NULL CHECK (quantity > 0),
     account_id VARCHAR NOT NULL,
     direction VARCHAR NOT NULL,
     created_at TIMESTAMP
